requirements.txt
asgiref==3.5.2
black==22.6.0
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
